wip: add config declarations
