+++ ckan/config/declaration/__init__.py
@@ -117,6 +117,8 @@ class Declaration:
         self.load_core_declaration()
         for plugin in p.PluginImplementations(p.IConfigDeclaration):
             plugin.declare_config_options(self, Key())
+
+        loader(self, "files")
         self._seal()
 
     def make_safe(self, config: "CKANConfig"):
