+++ ckan/lib/files/base.py
@@ -1,21 +1,20 @@
 from __future__ import annotations
 
 import dataclasses
-import os
+
 
 from time import time
-from typing import Any, Iterable, cast
+from typing import Any, Iterable
 
 import jwt
 import file_keeper as fk
 
 from typing_extensions import TypeAlias
 
-from ckan.common import config
+
 from ckan.config.declaration import Declaration
 from ckan.config.declaration.key import Key
 from ckan.lib.api_token import _get_algorithm, _get_secret
-from ckan.lib.helpers import url_for
 
 Uploader: TypeAlias = fk.Uploader
 Manager: TypeAlias = fk.Manager
@@ -23,115 +22,6 @@ Reader: TypeAlias = fk.Reader
 
 FileData: TypeAlias = fk.FileData
 MultipartData: TypeAlias = fk.MultipartData
-
-
-def collect_adapters() -> dict[str, type[Storage]]:
-    path = config["ckan.storage_path"]
-    result = {}
-
-    if path:
-        result["resources"] = make_storage(
-            "resources",
-            {
-                "type": "file_keeper:fs",
-                "path": os.path.join(path, "resources"),
-                "create_path": True,
-                "recursive": True,
-                "override_existing": True,
-                "location_transformers": ["safe_relative_path"],
-            },
-        )
-
-        for object_type in ["user", "group", "admin"]:
-            name = f"{object_type}_uploads"
-            result[name] = fk.make_storage(
-                name,
-                {
-                    "type": "file_keeper:fs",
-                    "path": os.path.join(path, "storage", "uploads", object_type),
-                    "create_path": True,
-                },
-            )
-
-    return result
-
-
-def collect_storages() -> dict[str, Storage]:
-    path = config["ckan.storage_path"]
-    result = {}
-
-    if path:
-        result["resources"] = make_storage(
-            "resources",
-            {
-                "type": "file_keeper:fs",
-                "path": os.path.join(path, "resources"),
-                "create_path": True,
-                "recursive": True,
-                "override_existing": True,
-                "location_transformers": ["safe_relative_path"],
-            },
-        )
-
-        for object_type in ["user", "group", "admin"]:
-            name = f"{object_type}_uploads"
-            result[name] = fk.make_storage(
-                name,
-                {
-                    "type": "file_keeper:fs",
-                    "path": os.path.join(path, "storage", "uploads", object_type),
-                    "create_path": True,
-                },
-            )
-
-    return result
-
-
-adapters = fk.Registry["type[Storage]"](collector=collect_adapters)
-storages = fk.Registry["Storage"](collector=collect_storages)
-
-
-def make_storage(name: str, settings: dict[str, Any]):
-    storage = fk.make_storage(name, settings)
-    if not isinstance(storage, Storage):
-        msg = "Does not extend ckan.lib.files.Storage"
-        raise fk.exc.InvalidStorageConfigurationError(name, msg)
-    return storage
-
-
-def get_storage(name: str | None = None) -> fk.Storage:
-    """Return existing storage instance.
-
-    Storages are initialized when plugin is loaded. As result, this function
-    always returns the same storage object for the given name.
-
-    If no name specified, default storage is returned.
-
-    Args:
-        name: name of the configured storage
-
-    Returns:
-        storage instance
-
-    Raises:
-        UnknownStorageError: storage with the given name is not configured
-
-    Example:
-        ```
-        default_storage = get_storage()
-        storage = get_storage("storage name")
-        ```
-
-    """
-    if name is None:
-        name = cast(str, config["ckan.files.default_storage"])
-
-    storage = storages.get(name)
-
-    if not storage:
-        raise fk.exc.UnknownStorageError(name)
-
-    return storage
 
 
 def is_supported_type(content_type: str, supported: Iterable[str]) -> bool:
@@ -191,6 +81,8 @@ class Storage(fk.Storage):
         return super().multipart_start(location, data, **kwargs)
 
     def temporal_link(self, data: FileData, /, **kwargs: Any) -> str:
+        from ckan.lib.helpers import url_for
+
         try:
             link = super().temporal_link(data, **kwargs)
         except fk.exc.UnsupportedOperationError:
